+++ src/main/java/insane96mcp/survivalreimagined/network/message/MessageTirednessSync.java
@@ -1,7 +1,7 @@
 package insane96mcp.survivalreimagined.network.message;
 
+import insane96mcp.survivalreimagined.module.sleeprespawn.feature.Tiredness;
 import insane96mcp.survivalreimagined.network.NetworkHelper;
-import insane96mcp.survivalreimagined.setup.Strings;
 import net.minecraft.network.FriendlyByteBuf;
 import net.minecraftforge.network.NetworkEvent;
 
@@ -23,7 +23,7 @@ public class MessageTirednessSync {
 	}
 
 	public static void handle(final MessageTirednessSync message, Supplier<NetworkEvent.Context> ctx) {
-		ctx.get().enqueueWork(() -> NetworkHelper.getSidedPlayer(ctx.get()).getPersistentData().putFloat(Strings.Tags.TIREDNESS, message.tiredness));
+		ctx.get().enqueueWork(() -> NetworkHelper.getSidedPlayer(ctx.get()).getPersistentData().putFloat(Tiredness.TIREDNESS_TAG, message.tiredness));
 		ctx.get().setPacketHandled(true);
 	}
 }