+++ changelog.md
@@ -1,7 +1,12 @@
 # Changelog
 
 ## Upcoming
+* Tiredness
+  * Increased tiredness required to sleep and effect
+  * Vigour on wake up changed (Vigour II for 8 minutes -> Vigour I for 20 minutes)
+  * Fixed tired applying continuously
 * Copper from campfire now requires 4 minutes instead of 5
+* Iridium tool durability increased by 10%
 
 ## Alpha 3.2.0
 * Enchantments